PasswordsDoNotMuchError, separate component for ChangePasswordForm, and internal use of copyOldUserDataToHistoryUserData implemented
