+++ apps/web/src/actions/error.ts
@@ -1,0 +1,6 @@
+export class PasswordsDoNotMatchError extends Error {
+    constructor() {
+      super();
+      Object.setPrototypeOf(this, PasswordsDoNotMatchError.prototype)
+    }
+}
