+++ apps/web/src/components/profile/base-information-form.tsx
@@ -1,10 +1,9 @@
 "use client";
 
 import { useTransition } from "react";
+import ChangePasswordForm from "./change-password-form";
 import { updateBaseInformationUsername } from "@/actions/profile";
-import { updateBaseInformationPassword } from "@/actions/profile";
 import { baseInfromationSchema } from "@/lib/schema/profile";
-import { passwordSchema } from "@/lib/schema/profile";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { Loader2Icon } from "lucide-react";
 import { useForm } from "react-hook-form";
@@ -34,27 +33,18 @@ interface Props {
 }
 
 export default function BaseInformationForm({ username, email, id }: Props) {
-    const [changeUsernameIsPending, startTransitionUsername] = useTransition();
-    const [changePasswordIsPending, startTransitionPassword] = useTransition();
+    const [changeIsPending, startTransition] = useTransition();
     const { toast } = useToast();
-    const formUsername = useForm<z.infer<typeof baseInfromationSchema>>({
+    const form = useForm<z.infer<typeof baseInfromationSchema>>({
         resolver: zodResolver(baseInfromationSchema),
         defaultValues: {
             username,
             email,
         },
     });
-    const formPassword = useForm<z.infer<typeof passwordSchema>>({
-        resolver: zodResolver(passwordSchema),
-        defaultValues: {
-            oldPassword: "",
-            newPassword: "",
-            newPasswordRepeat: "",
-        },
-    });
 
-    function onSubmitUsername(data: z.infer<typeof baseInfromationSchema>) {
-        startTransitionUsername(async () => {
+    function onSubmit(data: z.infer<typeof baseInfromationSchema>) {
+        startTransition(async () => {
             if (data.email !== email) {
                 return;
             }
@@ -74,41 +64,6 @@ export default function BaseInformationForm({ username, email, id }: Props) {
         });
     }
 
-    function onSubmitPassword(data: z.infer<typeof passwordSchema>) {
-        if (data.newPassword !== data.newPasswordRepeat) {
-            toast({
-                title: "Das neue Passwort stimmt nicht mit der Wiederholung überein",
-                description: "Dein Passwort konnte nicht geändert werden",
-                variant: "destructive",
-            });
-            return;
-        }
-
-        startTransitionPassword(async () => {
-            try {
-                await updateBaseInformationPassword(data, id);
-                toast({
-                    title: "Erfolgreich Passwort aktualisiert",
-                    description: "Dein Passwort wurde erfolgreich geändert",
-                });
-            } catch (e) {
-                if (e.message === "Passwords do not match") {
-                    toast({
-                        title: "Das aktuelle Passwort ist falsch",
-                        description: "Dein Passwort konnte nicht geändert werden",
-                        variant: "destructive",
-                    });
-                } else {
-                toast({
-                    title: "Fehler beim Aktualisieren des Passworts",
-                    description: "Dein Passwort konnte nicht geändert werden",
-                    variant: "destructive",
-                });
-                }
-            }
-        });
-    }
-
     return (
         <Card className="w-full">
             <CardHeader>
@@ -116,10 +71,10 @@ export default function BaseInformationForm({ username, email, id }: Props) {
                 <CardDescription>Deine persönlichen Daten</CardDescription>
             </CardHeader>
             <CardContent>
-                <Form {...formUsername}>
-                    <form className="grid grid-cols-2 gap-4" onSubmit={formUsername.handleSubmit(onSubmitUsername)}>
+                <Form {...form}>
+                    <form className="grid grid-cols-2 gap-4" onSubmit={form.handleSubmit(onSubmit)}>
                         <FormField
-                            control={formUsername.control}
+                            control={form.control}
                             name="username"
                             render={({ field }) => (
                                 <FormItem>
@@ -132,7 +87,7 @@ export default function BaseInformationForm({ username, email, id }: Props) {
                             )}
                         />
                         <FormField
-                            control={formUsername.control}
+                            control={form.control}
                             name="email"
                             render={({ field }) => (
                                 <FormItem>
@@ -145,69 +100,15 @@ export default function BaseInformationForm({ username, email, id }: Props) {
                             )}
                         />
                         <div className="col-span-2 flex flex-row justify-end">
-                            <Button disabled={changeUsernameIsPending} type="submit" value="username">
-                                {changeUsernameIsPending ? <Loader2Icon className="mr-2 h-4 w-4 animate-spin" /> : null}
+                            <Button disabled={changeIsPending} type="submit" value="username">
+                                {changeIsPending ? <Loader2Icon className="mr-2 h-4 w-4 animate-spin" /> : null}
                                 Speichern
                             </Button>
                         </div>
                     </form>
                 </Form>
             </CardContent>
-            <CardHeader>
-                <CardTitle>Passwort</CardTitle>
-                <CardDescription>Ändere dein Passwort</CardDescription>
-            </CardHeader>
-            <CardContent>
-                <Form {...formPassword}>
-                    <form className="flex flex-col gap-4" onSubmit={formPassword.handleSubmit(onSubmitPassword)}>
-                        <FormField
-                            control={formPassword.control}
-                            name="oldPassword"
-                            render={({ field }) => (
-                                <FormItem>
-                                    <FormLabel>Aktuelles Passwort</FormLabel>
-                                    <FormControl>
-                                        <Input placeholder="Aktuelles Passwort" type="password" {...field} />
-                                    </FormControl>
-                                    <FormMessage />
-                                </FormItem>
-                            )}
-                        />
-                        <FormField
-                            control={formPassword.control}
-                            name="newPassword"
-                            render={({ field }) => (
-                                <FormItem>
-                                    <FormLabel>Neues Passwort</FormLabel>
-                                    <FormControl>
-                                        <Input placeholder="Neues Passwort" type="password" {...field} />
-                                    </FormControl>
-                                    <FormMessage />
-                                </FormItem>
-                            )}
-                        />
-                        <FormField
-                            control={formPassword.control}
-                            name="newPasswordRepeat"
-                            render={({ field }) => (
-                                <FormItem>
-                                    <FormLabel>Passwort Wiederholen</FormLabel>
-                                    <FormControl>
-                                        <Input placeholder="Passwort Wiederholen" type="password" {...field} />
-                                    </FormControl>
-                                    <FormMessage />
-                                </FormItem>
-                            )}
-                        />
-                        <div className="col-span-2 flex flex-row justify-end">
-                            <Button disabled={changePasswordIsPending} type="submit" value="password">
-                                {changePasswordIsPending ? <Loader2Icon className="mr-2 h-4 w-4 animate-spin" /> : null}
-                                Speichern
-                            </Button>
-                        </div>
-                    </form>
-                </Form>
-            </CardContent>
+            <ChangePasswordForm id={id}/>
         </Card>
     );
 }
