+++ packages/db/src/query/user.ts
@@ -137,10 +137,11 @@ type UpdateUserData = {
 };
 
 export async function updateUserData(data: UpdateUserData, id: number) {
+    await copyOldUserDataToHistoryUserData(id);
     return await db.update(userData).set(data).where(eq(userData.userId, id));
 }
 
-export async function copyOldUserDataToHistoryUserData(userId: number) {
+async function copyOldUserDataToHistoryUserData(userId: number) {
     return db.transaction(async (trx) => {
         const oldUserData = await getUserDataByUserId(userId);
         if (!oldUserData) {
